requirements.txt
streamlit==1.4.0
PyPDF2==1.26.0
langchain==0.1.2
python-dotenv==0.15.0
openai==0.27.0
faiss==1.7.1
